+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Version 3.0.27
-- Proof of concept for improved editing of powers.  Currently only RKA is using the improved editing.
+- Proof of concept for improved editing of powers.  Open item, click on Sheet in header, then select Herosystem6eItem2Sheet to preview.
+- Improved actor migration to update power modifiers. [#287](https://github.com/dmdorman/hero6e-foundryvtt/issues/287)
 
 # Version 3.0.26
 - Testing workflow to publish to FoundryVTT.